feat(worktrees): manage parallel worktrees from a canvas toolbar drop-up
Promote parallel-worktree tooling out of the sidebar and into a drop-up next
to the Hand tool, then retire the sidebar tab.

The drop-up (WorktreeToolbarMenu) lets you, per worktree: focus its spatial
lens, see live git status + PR state + how many terminals/agents are open on
the canvas, open a terminal or Cate agent bound to it (click = here, drag =
drop anywhere), recolor/rename inline, reach the full publish/PR/update/merge/
discard menu via the overflow, plus create a new worktree, clean up orphans,
and init git when the folder isn't a repo yet.

Shared, deduplicated logic:
- useParallelWork — the worktree verbs + native context menu
- useWorktreeActions — create / PR-checkout
- useWorktreeStatuses — per-worktree dirty status + gh PR lookups + humanStatus
- CreateWorktreeForm — the "what are you working on?" form (inline/flat variants)

Remove the Parallel Work sidebar tab and its SidebarView; metadata sync already
runs app-wide via worktreeSync/useProcessMonitor, and per-worktree status/PR is
now fetched on demand only while the drop-up is open.
