+++ src/renderer/canvas/CanvasToolbar.tsx
@@ -19,6 +19,7 @@ import {
 } from '@phosphor-icons/react'
 import { CateLogo } from '../ui/CateLogo'
 import Minimap from './Minimap'
+import WorktreeToolbarMenu from './WorktreeToolbarMenu'
 import { useCanvasStoreApi } from '../stores/CanvasStoreContext'
 import { useUIStore } from '../stores/uiStore'
 import { useUIStateStore } from '../stores/uiStateStore'
@@ -34,6 +35,8 @@ const loadMinimapCorner = (): MinimapCorner => useUIStateStore.getState().minima
 
 interface CanvasToolbarProps {
   canvasPanelId: string
+  workspaceId: string
+  rootPath: string
   zoom: number
   onNewTerminal: () => void
   onNewBrowser: () => void
@@ -187,6 +190,8 @@ const ModeButton: React.FC<{
 
 const CanvasToolbar: React.FC<CanvasToolbarProps> = ({
   canvasPanelId,
+  workspaceId,
+  rootPath,
   zoom,
   onNewTerminal,
   onNewBrowser,
@@ -277,6 +282,14 @@ const CanvasToolbar: React.FC<CanvasToolbarProps> = ({
             >
               <Hand size={18} />
             </ModeButton>
+
+            {/* Parallel worktrees — drop-up: focus a worktree's spatial lens,
+                open a terminal in one, or start a new parallel branch. */}
+            <WorktreeToolbarMenu
+              canvasPanelId={canvasPanelId}
+              workspaceId={workspaceId}
+              rootPath={rootPath}
+            />
 
             {/* Divider */}
             <div className="w-px h-5 bg-surface-5 mx-1" />
