+++ src/renderer/sidebar/Sidebar.tsx
@@ -3,7 +3,6 @@ import { ProjectList } from './ProjectList'
 import { FileExplorer } from './FileExplorer'
 import { SearchView } from './SearchView'
 import { SourceControlView } from './SourceControlView'
-import { ParallelWorkTab } from './ParallelWorkTab'
 import { UpdateButton } from './UpdateButton'
 import { useAppStore } from '../stores/appStore'
 import { useUIStore, useSidebarLayout } from '../stores/uiStore'
@@ -11,7 +10,6 @@ import type { SidebarView, SidebarSide } from '../stores/uiStore'
 import {
   FolderOpen,
   GitBranch,
-  ArrowsSplit,
   Stack,
   Gear,
   MagnifyingGlass,
@@ -29,7 +27,6 @@ const VIEW_META: Record<SidebarView, { icon: PhosphorIcon; title: string }> = {
   explorer: { icon: FolderOpen, title: 'Explorer' },
   search: { icon: MagnifyingGlass, title: 'Search' },
   git: { icon: GitBranch, title: 'Source Control' },
-  parallelWork: { icon: ArrowsSplit, title: 'Parallel Work' },
 }
 
 // ---------------------------------------------------------------------------
@@ -70,8 +67,6 @@ const SidebarViewContent: React.FC<{ view: SidebarView; rootPath: string }> = ({
       return <SearchView rootPath={rootPath} workspaceId={selectedWorkspaceId} />
     case 'git':
       return <SourceControlView rootPath={rootPath} />
-    case 'parallelWork':
-      return <ParallelWorkTab rootPath={rootPath} />
     default:
       return null
   }
