+++ src/shared/types.ts
@@ -1084,7 +1084,7 @@ export const FILE_EXCLUSIONS: string[] = [
 ]
 
 /** A sidebar view (left/right rail tabs). */
-export type SidebarView = 'workspaces' | 'explorer' | 'git' | 'parallelWork' | 'search'
+export type SidebarView = 'workspaces' | 'explorer' | 'git' | 'search'
 
 /** Which sidebar views live in the left vs. right rail. Persisted in settings. */
 export interface SidebarLayout {
@@ -1297,7 +1297,7 @@ export const DEFAULT_SETTINGS: AppSettings = {
   // Layout — keep in sync with the sidebar's default arrangement.
   sidebarLayout: {
     left: ['workspaces', 'explorer', 'search'],
-    right: ['git', 'parallelWork'],
+    right: ['git'],
   },
 }
